pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lakebase-utils"
version = "0.1.0"
description = "Utility package for Databricks Lakebase Autoscaling operations"
requires-python = ">=3.10"
dependencies = [
    "databricks-sdk>=0.30",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
]

[tool.hatch.build.targets.wheel]
packages = ["src/lakebase_utils"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
